Orography-aware 2D convolution (#798)
* Correct handling of geometry fields

* Bugfix

* Use fields metadata

* Trigger tests

* Following Francois suggestion, not done yet

* Simplify BUMP interface, move all sources toghether

* New attempt

* Passing fields metadata to variables metadata

* Remove routine name

* Add metadata in random()

* Add test with BUMP instances using variables with different number of levels

* WIP

* Correct abort

* Metadata specified in the yaml for each block

* Update for latest merge

* Remove halo exchange

* Minor cleaning

* Better block name

* Move to external conf section

* Improved yaml options, complete testing

* Fix coding norms

* Remove useless halo exchange

* Fixes

* Unrelated fixes in Fields.cc to remove compiler warnings

---------

+++ src/saber/blocks/SaberOuterBlockBase.cc
@@ -107,9 +107,9 @@ void SaberOuterBlockBase::adjointTest(const oops::GeometryData & outerGeometryDa
     auto ghostView = atlas::array::make_view<int, 1>(outerGeometryData.functionSpace().ghost());
     for (const auto & var : outerVars.variables()) {
       auto view = atlas::array::make_view<double, 2>(outerFset[var]);
-      for (size_t jn = 0; jn < view.shape(0); ++jn) {
+      for (int jn = 0; jn < view.shape(0); ++jn) {
         if (ghostView(jn) == 1) {
-          for (size_t jl = 0; jl < view.shape(1); ++jl) {
+          for (int jl = 0; jl < view.shape(1); ++jl) {
             view(jn, jl) = 0.0;
           }
         }
